[Refactor] Moved Box and PullRefreshIndicator to NewsListScreenContent
