--- news/news_ui/src/main/java/com/example/news_ui/screen/list/NewsUiState.kt
@@ -1,10 +1,0 @@
-package com.example.news_ui.screen.list
-
-
-data class NewsUiState(
-    val id: String,
-    val title: String,
-    val writtenAt: String,
-    val section: String,
-    val onClick: (id: String) -> Unit
-)
